backend/pyproject.toml
[project]
name = "CoSMIC"
version = "0.1.0"
description = "Cognitive System of Machine Intelligent Computing"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.129.0",
    # NOTE: Use 'psycopg[binary]' version on pord.
    "psycopg>=3.3.3",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.37",
    # TODO: Installed for now to get CoSMIC file working.
    "python-box>=7.4.1",
    "torch>=2.10.0",
    "ollama>=0.6.1",
    "transformers[torch]>=5.3.0",
    "peft>=0.18.1",
    "openai>=2.26.0",
    "pytz>=2026.1.post1",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "chess>=1.11.2",
    "stockfish>=4.0.8",
    "pandas>=3.0.1",
]
